pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nymble-relay"
version = "0.1.0"
description = "Headless text-to-keystroke relay daemon"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "websockets>=12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
hid = ["pyserial>=3.5"]
all = ["pyserial>=3.5"]

[project.scripts]
nymble-relay = "nymble_relay.__main__:main"
